refactor(settings): extract SettingsStatsGrid + revert iconColor + move default models
- Extract SettingsStatsGrid (presentational, takes columns prop) out of
  SettingsLayoutItemsStats. Reuse it on the AI Overview tab — the
  bespoke StatTile/StyledStatCard/StyledStatsGrid I had there gets
  deleted. SettingsLayoutItemsStats now just supplies its own column
  data and renders SettingsStatsGrid.
- Move the Default model section (Smart / Fast pickers) off the AI
  Overview tab and onto the Models tab as its first section, above
  'Available models'. Overview keeps At-a-glance + MCP signpost +
  Workspace Instructions.
- Revert all iconColor work on the settings nav: strip the iconColor
  field on SettingsNavigationItem, drop iconColor={item.iconColor}
  from SettingsNavigationDrawerItem's two NavigationDrawerItem
  renders, and remove every per-item iconColor: '...' on
  useSettingsNavigationItems. The settings menu icons now render the
  same monochrome grey-on-grey tiles main does. The tinted-tile
  treatment is parked for a separate PR.
