+++ packages/twenty-front/src/modules/settings/components/SettingsNavigationDrawerItem.tsx
@@ -39,7 +39,6 @@ export const SettingsNavigationDrawerItem = ({
           label={item.label}
           to={href}
           Icon={item.Icon}
-          iconColor={item.iconColor}
           active={isActive}
           modifier={item.modifier}
           onClick={item.onClick}
@@ -55,7 +54,6 @@ export const SettingsNavigationDrawerItem = ({
       label={item.label}
       to={href || undefined}
       Icon={item.Icon}
-      iconColor={item.iconColor}
       active={isActive}
       modifier={item.modifier}
       onClick={item.onClick}
