+++ packages/twenty-front/src/pages/settings/ai/components/SettingsAiModelsTab.tsx
@@ -3,14 +3,31 @@ import { useContext, useState } from 'react';
 
 import { useWorkspaceAiModelAvailability } from '@/ai/hooks/useWorkspaceAiModelAvailability';
 import { currentWorkspaceState } from '@/auth/states/currentWorkspaceState';
+import { aiModelsState } from '@/client-config/states/aiModelsState';
 import { SettingsAiModelsTable } from '@/settings/ai/components/SettingsAiModelsTable';
+import { getDataResidencyDisplay } from '@/settings/ai/utils/getDataResidencyDisplay';
+import { getModelIcon } from '@/settings/ai/utils/getModelIcon';
 import { SettingsCard } from '@/settings/components/SettingsCard';
+import { SettingsOptionCardContentSelect } from '@/settings/components/SettingsOptions/SettingsOptionCardContentSelect';
 import { SettingsOptionCardContentToggle } from '@/settings/components/SettingsOptions/SettingsOptionCardContentToggle';
 import { useSnackBar } from '@/ui/feedback/snack-bar-manager/hooks/useSnackBar';
+import { Select } from '@/ui/input/components/Select';
+import { GenericDropdownContentWidth } from '@/ui/layout/dropdown/constants/GenericDropdownContentWidth';
 import { useAtomState } from '@/ui/utilities/state/jotai/hooks/useAtomState';
+import { useAtomStateValue } from '@/ui/utilities/state/jotai/hooks/useAtomStateValue';
 import { useMutation, useQuery } from '@apollo/client/react';
 import { t } from '@lingui/core/macro';
-import { H2Title, IconPrompt, IconStar } from 'twenty-ui/display';
+import {
+  AUTO_SELECT_FAST_MODEL_ID,
+  AUTO_SELECT_SMART_MODEL_ID,
+} from 'twenty-shared/constants';
+import {
+  H2Title,
+  IconBolt,
+  IconBrain,
+  IconPrompt,
+  IconStar,
+} from 'twenty-ui/display';
 import { SearchInput } from 'twenty-ui/input';
 import { Card, Section } from 'twenty-ui/layout';
 import { UndecoratedLink } from 'twenty-ui/navigation';
@@ -30,9 +47,6 @@ const StyledCustomModelsContainer = styled.div`
   padding-top: ${themeCssVariables.spacing[4]};
 `;
 
-// Default model pickers (Smart / Fast) moved to SettingsAiOverviewTab —
-// this tab is the catalog (Use-best-models toggle + searchable list of
-// available models) plus the read-only System Prompt link card.
 export const SettingsAiModelsTab = () => {
   const { theme } = useContext(ThemeContext);
   const { enqueueErrorSnackBar } = useSnackBar();
@@ -42,6 +56,7 @@ export const SettingsAiModelsTab = () => {
   const [updateWorkspace] = useMutation(UpdateWorkspaceDocument);
   const [searchQuery, setSearchQuery] = useState('');
   const { data: previewData } = useQuery(GetAiSystemPromptPreviewDocument);
+  const aiModels = useAtomStateValue(aiModelsState);
 
   const systemPromptTokenCount =
     previewData?.getAiSystemPromptPreview.estimatedTokenCount;
@@ -52,8 +67,58 @@ export const SettingsAiModelsTab = () => {
       )} tokens)`
     : t`Read the system prompts to understand how the AI works`;
 
-  const { useRecommendedModels, realModels } =
+  const { useRecommendedModels, realModels, enabledModels } =
     useWorkspaceAiModelAvailability();
+
+  // Default model pickers (Smart / Fast) — pinned "auto-select" entry is the
+  // "Best" option for each tier, options pull from the workspace-enabled set.
+  const currentSmartModel = currentWorkspace?.smartModel;
+  const currentFastModel = currentWorkspace?.fastModel;
+
+  const buildPinnedOption = (autoSelectModelId: string) => {
+    const autoSelectEntry = aiModels.find(
+      (model) => model.modelId === autoSelectModelId,
+    );
+    if (!autoSelectEntry) return undefined;
+    return {
+      value: autoSelectModelId,
+      label: autoSelectEntry.label,
+      Icon: getModelIcon(
+        autoSelectEntry.modelFamily,
+        autoSelectEntry.providerName,
+      ),
+      contextualText: t`Best`,
+    };
+  };
+
+  const smartPinnedOption = buildPinnedOption(AUTO_SELECT_SMART_MODEL_ID);
+  const fastPinnedOption = buildPinnedOption(AUTO_SELECT_FAST_MODEL_ID);
+
+  const modelOptions = enabledModels.map((model) => {
+    const residencyFlag = model.dataResidency
+      ? ` ${getDataResidencyDisplay(model.dataResidency)}`
+      : '';
+    return {
+      value: model.modelId,
+      label: `${model.label}${residencyFlag}`,
+      Icon: getModelIcon(model.modelFamily, model.providerName),
+    };
+  });
+
+  const handleModelFieldChange = async (
+    field: 'smartModel' | 'fastModel',
+    value: string,
+  ) => {
+    if (!currentWorkspace?.id) return;
+    const previousValue = currentWorkspace[field];
+    try {
+      setCurrentWorkspace({ ...currentWorkspace, [field]: value });
+      await updateWorkspace({ variables: { input: { [field]: value } } });
+    } catch {
+      setCurrentWorkspace({ ...currentWorkspace, [field]: previousValue });
+      enqueueErrorSnackBar({ message: t`Failed to update model` });
+    }
+  };
 
   const enabledModelIdSet = new Set(currentWorkspace?.enabledAiModelIds ?? []);
 
@@ -154,6 +219,45 @@ export const SettingsAiModelsTab = () => {
 
   return (
     <>
+      <Section>
+        <H2Title
+          title={t`Default model`}
+          description={t`The default AI model used for chats, agents, and workflows`}
+        />
+        <Card rounded>
+          <SettingsOptionCardContentSelect
+            Icon={IconBrain}
+            title={t`Smart Model`}
+            description={t`Used for chats, agents, and complex reasoning`}
+          >
+            <Select
+              dropdownId="models-tab-smart-model-select"
+              value={currentSmartModel}
+              onChange={(value) => handleModelFieldChange('smartModel', value)}
+              options={modelOptions}
+              pinnedOption={smartPinnedOption}
+              selectSizeVariant="small"
+              dropdownWidth={GenericDropdownContentWidth.ExtraLarge}
+            />
+          </SettingsOptionCardContentSelect>
+          <SettingsOptionCardContentSelect
+            Icon={IconBolt}
+            title={t`Fast Model`}
+            description={t`Used for lightweight tasks like title generation`}
+          >
+            <Select
+              dropdownId="models-tab-fast-model-select"
+              value={currentFastModel}
+              onChange={(value) => handleModelFieldChange('fastModel', value)}
+              options={modelOptions}
+              pinnedOption={fastPinnedOption}
+              selectSizeVariant="small"
+              dropdownWidth={GenericDropdownContentWidth.ExtraLarge}
+            />
+          </SettingsOptionCardContentSelect>
+        </Card>
+      </Section>
+
       <Section>
         <H2Title
           title={t`Available models`}
